fix review

+++ native/cocos/audio/android/AudioDecoder.h
@@ -48,7 +48,7 @@ protected:
     static size_t fileRead(void *ptr, size_t size, size_t nmemb, void *datasource);
     static int fileSeek(void *datasource, int64_t offset, int whence);
     static int fileClose(void *datasource);
-    static int32_t fileTell(void *datasource);
+    static long fileTell(void *datasource);
 
     ccstd::string _url;
     PcmData _result;
